move base device into base_device.py
